Move the scaling code, especially "best fit", out of the OpenGL pipe and into the ViewEditor.  This ensures that the same logic is being used for all scaling operation, and take business logic out of the renderer.  The renderer is becoming more and more exclusively a rendering engine which is directed by the ViewEditor.
Since Java favors double, most calculations are now computed as double, with
conversion to float only when saving in Point2D objects.  This should reduce
the number of type conversions through the rending cycle.

+++ DepanApp/prod/src/com/google/devtools/depan/eclipse/visualization/ogl/RendererChangeListener.java
@@ -69,5 +69,5 @@ public interface RendererChangeListener {
    * Notify the receiver that the position all of currently selected nodes
    * should be adjusted by the relative amounts.
    */
-  public void selectionMoved(float x, float y);
+  public void selectionMoved(double x, double y);
 }
